* V5-7-patches:
  MinGW: Ensure that pskill / pslist test is run before tests
  Win32: Make strcasecmp() declaration available to the compiler
  Win32: Enable _WINSOCK_DEPRECATED_NO_WARNINGS
  Cygwin: getnetent() / getnetbyaddr() build fix
  UCD-SNMP MIB: Fix a link failure on systems without regex support
  UCD-SNMP MIB: Avoid integer overflows
  snmpd: Rename VIEWSTATUS into VACMVIEWSTATUS
  CHANGES: BUG: 2737: Quote service executable path
  Win32: Add support for OpenSSL >= 1.1.0
  Win32: Add Visual Studio 2015 version number
  Win32: Synchronize win32/net-snmp/net-snmp-config.h and .in
  snmplib/winservice.c: Suppress a compiler warning
  snmplib: Add missing NETSNMP_IMPORT declarations
  snmplib: Fix asn_parse_nlength() prototype
  snmplib: Unify printf format attribute definitions

+++ agent/mibgroup/ucd-snmp/memory.c
@@ -59,7 +59,7 @@ handle_memory(netsnmp_mib_handler *handler,
                 netsnmp_request_info *requests)
 {
     netsnmp_memory_info *mem_info;
-    unsigned long val;
+    unsigned long long val;
     struct counter64 c64;
     int type = ASN_INTEGER;
     char buf[1024];
